requirements.txt
fastapi
uvicorn[standard]
httpx
pyjwt[crypto]
pytest
prometheus-fastapi-instrumentator